+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+IMPROVEMENTS:
+
+* resource/security/resource_artifactory_permission_target: Add support for `distribute` permission for `release_bundle`.
+
 ## 6.9.2 (June 7, 2022). Tested on Artifactory 7.38.10
 
 BUG FIXES: